Update to ruby 2.7.6 and Rails 6.0.3.6
The rails@6.0.2 gem depended on marcel@0.3.1, which has been yanked from the package registry because of license issues. Updated to the closest version rails@6.0.3.6 which relies on marcei@1.0.0, which fixes the issue.

Updated Ruby to 2.7.6. After having issues installing 2.7.0 on an ARM Mac via ruby-build, installed Ruby 2.7.6 with Homebrew. Ruby 2.7.0 was not available on Homebrew.

Node 14 is required by node-sass 4.14.1, one of the dependencies.